requirements.txt
alembic==1.3.1
Babel==2.7.0
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-translate==2.0.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
guess-language-spirit==0.5.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
